config_local: store TLF path along with sync config
To truly work offline, we'll need to be able to tell the service that
we want to resolve a given path name with offline availability, even
before we know the TLF ID.  So we need to track the synced path names
in the sync config DB as well.

This will be used in future work to inform set offline availability
settings on service RPCs.

Issue: KBFS-3918